requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.21.0
python-dotenv>=1.0.1
pydantic>=1.10
requests>=2.31.0
pydantic[email]>=1.10